+++ changelog.d/23480_dedupe_transform_time_settings.feature.md
@@ -1,3 +1,3 @@
-Adds `timed` configuration for `dedupe` transform, to enable configuring `max_age` for items in the deduplication cache, helping make a distinction between real duplicates and duplication that is expected in the incoming data over a longer period of time.
+Adds `time_settings` configuration for `dedupe` transform, to enable configuring `max_age` for items in the deduplication cache, helping make a distinction between real duplicates and duplication that is expected in the incoming data over a longer period of time.
 
 authors: esensar Quad9DNS
